+++ gallery_generator/templates/tag.html
@@ -1,0 +1,6 @@
+{% extends "base.ext.html" %}
+
+{% block page_content %}
+    {% for picture in pictures %}
+    {% endfor %}
+{% endblock %}
